project/course_planner/requirements.txt
google-genai
streamlit
pypdf2
python-dotenv
openpyxl
ratemyprofessors-client
streamlit-authenticator
sqlite-vec
bcrypt
pytest
